test(local-inference): rename fake to SidecarConnection.fake.ts for clarity
The test double sat next to production code under a plain .ts name with no
test-suffix convention, reading as production. Rename to SidecarConnection.fake.ts
(git mv preserves history) so the trio SidecarConnection.ts / .fake.ts / .test.ts
is self-documenting. Test-only: no app code imports it; production bundle unaffected.
