+++ src/lib/local-inference/native/SidecarConnection.fake.ts
